+++ app/handlers/newflow/educator_signup/complete_profile.rb
@@ -1,6 +1,7 @@
 module Newflow
   module EducatorSignup
     class CompleteProfile
+
       OTHER = 'other'
       INSTRUCTOR = 'instructor'
       AS_PRIMARY = 'as_primary'
@@ -32,11 +33,16 @@ module Newflow
       end
 
       def handle
-        educator_params_check
+        check_params
         return if errors?
 
         caller.update(
-          role: signup_params.educator_specific_role
+          role: signup_params.educator_specific_role,
+          other_role_name: other_role_name,
+          using_openstax_how: signup_params.using_openstax_how,
+          who_chooses_books: signup_params.how_are_books_chosen,
+          how_many_students: signup_params.num_students_per_semester_taught,
+          which_books: books_or_subjects
         )
         transfer_errors_from(caller, {type: :verbatim}, :fail_if_errors)
 
@@ -45,7 +51,26 @@ module Newflow
 
       private #################
 
-      def educator_params_check
+      def update_salesforce_lead
+        UpdateSalesforceLead.perform_later(user: caller)
+      end
+
+      def other_role_name
+        signup_params.educator_specific_role == OTHER ? signup_params.other_role_name.strip : nil
+      end
+
+      def books_or_subjects
+        subjects = signup_params.subjects_of_interest&.reject(&:empty?)
+        titles = signup_params.books_used&.reject(&:empty?)
+
+        if subjects&.any?
+          subjects.join(';')
+        elsif titles&.any?
+          titles.join(';')
+        end
+      end
+
+      def check_params
         if signup_params.educator_specific_role.strip.downcase == OTHER &&
           signup_params.other_role_name.blank?
 
@@ -82,44 +107,6 @@ module Newflow
         )
       end
 
-      def update_salesforce_lead
-        # if Settings::Salesforce.push_leads_enabled
-        #   UpdateSalesforceLead.perform_later(
-        #     user: caller,
-        #     num_students: signup_params.num_students_per_semester_taught,
-        #     using_openstax: signup_params.using_openstax_how,
-        #     subject: books_or_subjects,
-
-        #     role: kind_of_educator,
-        #     # school: signup_params.self_reported_school, # TODO
-
-        #     # TODO: salesforce must be ready to accept these before we try to push them
-        #     # more_specific_educator_role: signup_params.educator_specific_role || signup_params.other_role_name,
-        #     # titles: signup_params.books_used.join(';'),
-        #     # salesforce_lead_id: caller.salesforce_lead_id,
-        #     # salesforce_contact_id: caller.salesforce_contact_id,
-
-        #     # do not set source_application because this faculty access endpoint does
-        #     # not have a strong indication of where the user is coming from
-        #     source_application: nil
-        #   )
-        # end
-      end
-
-      def kind_of_educator
-        signup_params.educator_specific_role == OTHER ? signup_params.other_role_name : signup_params.educator_specific_role
-      end
-
-      def books_or_subjects
-        subjects = signup_params.subjects_of_interest&.reject(&:empty?)
-        titles = signup_params.books_used&.reject(&:empty?)
-
-        if subjects&.any?
-          return subjects.join(';')
-        elsif titles&.any?
-          return titles.join(';')
-        end
-      end
     end
   end
 end
