+++ db/migrate/20200701233800_add_fields_for_signup_step_4_to_users.rb
@@ -1,0 +1,9 @@
+class AddFieldsForSignupStep4ToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :other_role_name, :string
+    add_column :users, :how_many_students, :string
+    add_column :users, :which_books, :string
+    add_column :users, :who_chooses_books, :string
+    add_column :users, :using_openstax_how, :integer
+  end
+end
