+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_06_18_191456) do
+ActiveRecord::Schema.define(version: 2020_07_01_233800) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "citext"
@@ -355,6 +355,11 @@ ActiveRecord::Schema.define(version: 2020_06_18_191456) do
     t.string "sheerid_verification_id"
     t.boolean "opt_out_of_cookies", default: false
     t.string "salesforce_lead_id"
+    t.string "other_role_name"
+    t.string "how_many_students"
+    t.string "which_books"
+    t.string "who_chooses_books"
+    t.integer "using_openstax_how"
     t.index "lower((first_name)::text)", name: "index_users_on_first_name"
     t.index "lower((last_name)::text)", name: "index_users_on_last_name"
     t.index "lower((username)::text)", name: "index_users_on_username_case_insensitive"
