+++ spec/cassettes/Newflow/EducatorSignup/CreateSalesforceLead/sf_setup.yml
@@ -21,7 +21,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Tue, 23 Jun 2020 20:02:43 GMT
+      - Fri, 03 Jul 2020 01:14:05 GMT
       Strict-Transport-Security:
       - max-age=31536000; includeSubDomains
       X-Content-Type-Options:
@@ -31,8 +31,8 @@ http_interactions:
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=gDPSl7WMEeqh6ztQ7cmS5Q; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 20:02:43 GMT; Max-Age=31536000
+      - BrowserId=fQRJXLzKEeq08wVW1YDU4Q; domain=.salesforce.com; path=/; expires=Sat,
+        03-Jul-2021 01:14:05 GMT; Max-Age=31536000
       Expires:
       - Thu, 01 Jan 1970 00:00:00 GMT
       X-Readonlymode:
@@ -45,7 +45,7 @@ http_interactions:
       - chunked
     body:
       encoding: ASCII-8BIT
-      string: '{"access_token":"<salesforce_access_token>","instance_url":"<salesforce_instance_url>","id":"<salesforce_id>","token_type":"Bearer","issued_at":"1592942564145","signature":"<salesforce_signature>"}'
+      string: '{"access_token":"<salesforce_access_token>","instance_url":"<salesforce_instance_url>","id":"<salesforce_id>","token_type":"Bearer","issued_at":"1593738846129","signature":"<salesforce_signature>"}'
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 20:02:44 GMT
+  recorded_at: Fri, 03 Jul 2020 01:14:06 GMT
 recorded_with: VCR 3.0.3
