+++ spec/cassettes/Newflow_EducatorSignup_CreateSalesforceLead/works_on_the_happy_path.yml
@@ -5,8 +5,8 @@ http_interactions:
     uri: "<salesforce_instance_url_lower>/services/data/v37.0/sobjects/Lead"
     body:
       encoding: UTF-8
-      string: '{"FirstName":"Maddie","LastName":"Kulas","Company":"not yet known","Email":"f@f.com","LeadSource":"OSC
-        Faculty","Newsletter__c":true,"Newsletter_Opt_In__c":true,"OS_Accounts_ID__c":1,"Application_Source__c":"Tutor
+      string: '{"FirstName":"Armanda","LastName":"Bogan","Company":"not yet known","Email":"f@f.com","LeadSource":"OSC
+        Faculty","Newsletter__c":true,"Newsletter_Opt_In__c":true,"OS_Accounts_ID__c":1,"accounts_uuid_c__c":"ac80c504-7001-434b-94db-b25bafcef0d1","Application_Source__c":"Tutor
         Signup","Role__c":"instructor"}'
     headers:
       User-Agent:
@@ -25,7 +25,7 @@ http_interactions:
       message: Created
     headers:
       Date:
-      - Tue, 23 Jun 2020 05:48:41 GMT
+      - Fri, 03 Jul 2020 01:14:07 GMT
       Strict-Transport-Security:
       - max-age=31536002; includeSubDomains
       Public-Key-Pins-Report-Only:
@@ -37,20 +37,20 @@ http_interactions:
       X-Robots-Tag:
       - none
       X-B3-Traceid:
-      - ce847e86afd4f082
+      - e0ea7867db03cfb0
       X-B3-Spanid:
-      - ce847e86afd4f082
+      - e0ea7867db03cfb0
       X-B3-Sampled:
       - '0'
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=MVyJfLUVEeqgMm0MFgS2ZA; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 05:48:41 GMT; Max-Age=31536000
+      - BrowserId=fkQcarzKEeqmxjvGPZToEg; domain=.salesforce.com; path=/; expires=Sat,
+        03-Jul-2021 01:14:07 GMT; Max-Age=31536000
       Sforce-Limit-Info:
-      - api-usage=5120/51000
+      - api-usage=2686/51000
       Location:
-      - "/services/data/v37.0/sobjects/Lead/00QR000000J6NSzMAN"
+      - "/services/data/v37.0/sobjects/Lead/00QR000000JNXLTMA5"
       Content-Type:
       - application/json;charset=UTF-8
       Vary:
@@ -59,12 +59,12 @@ http_interactions:
       - chunked
     body:
       encoding: ASCII-8BIT
-      string: '{"id":"00QR000000J6NSzMAN","success":true,"errors":[]}'
+      string: '{"id":"00QR000000JNXLTMA5","success":true,"errors":[]}'
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 05:48:44 GMT
+  recorded_at: Fri, 03 Jul 2020 01:14:10 GMT
 - request:
     method: get
-    uri: "<salesforce_instance_url_lower>/services/data/v37.0/query?q=SELECT%20Id,%20Name,%20FirstName,%20LastName,%20Salutation,%20Subject__c,%20Company,%20Phone,%20Website,%20Status,%20Email,%20LeadSource,%20Newsletter__c,%20Newsletter_Opt_In__c,%20Adoption_Status__c,%20Number_of_Students__c,%20OS_Accounts_ID__c,%20accounts_uuid_c__c,%20Application_Source__c,%20Role__c%20FROM%20Lead%20WHERE%20(Id%20=%20%2700QR000000J6NSzMAN%27)%20LIMIT%201"
+    uri: "<salesforce_instance_url_lower>/services/data/v37.0/query?q=SELECT%20Id,%20Name,%20FirstName,%20LastName,%20Salutation,%20Subject__c,%20Company,%20Phone,%20Website,%20Status,%20Email,%20LeadSource,%20Newsletter__c,%20Newsletter_Opt_In__c,%20Adoption_Status__c,%20Number_of_Students__c,%20OS_Accounts_ID__c,%20accounts_uuid_c__c,%20Application_Source__c,%20Role__c,%20who_chooses_books__c,%20FV_Final__c%20FROM%20Lead%20WHERE%20(Id%20=%20%2700QR000000JNXLTMA5%27)%20LIMIT%201"
     body:
       encoding: US-ASCII
       string: ''
@@ -83,7 +83,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Tue, 23 Jun 2020 05:48:44 GMT
+      - Fri, 03 Jul 2020 01:14:11 GMT
       Strict-Transport-Security:
       - max-age=31536002; includeSubDomains
       Public-Key-Pins-Report-Only:
@@ -95,18 +95,18 @@ http_interactions:
       X-Robots-Tag:
       - none
       X-B3-Traceid:
-      - 5620300347dd727f
+      - a605c7c4366f7834
       X-B3-Spanid:
-      - 5620300347dd727f
+      - a605c7c4366f7834
       X-B3-Sampled:
       - '0'
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=M2j6nLUVEeqtKwkbNZArmQ; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 05:48:44 GMT; Max-Age=31536000
+      - BrowserId=gFSFurzKEeqFmrnUwSpwPw; domain=.salesforce.com; path=/; expires=Sat,
+        03-Jul-2021 01:14:11 GMT; Max-Age=31536000
       Sforce-Limit-Info:
-      - api-usage=5120/51000
+      - api-usage=2686/51000
       Content-Type:
       - application/json;charset=UTF-8
       Vary:
@@ -115,12 +115,12 @@ http_interactions:
       - chunked
     body:
       encoding: ASCII-8BIT
-      string: '{"totalSize":1,"done":true,"records":[{"attributes":{"type":"Lead","url":"/services/data/v37.0/sobjects/Lead/00QR000000J6NSzMAN"},"Id":"00QR000000J6NSzMAN","Name":"Maddie
-        Kulas","FirstName":"Maddie","LastName":"Kulas","Salutation":null,"Subject__c":null,"Company":"not
+      string: '{"totalSize":1,"done":true,"records":[{"attributes":{"type":"Lead","url":"/services/data/v37.0/sobjects/Lead/00QR000000JNXLTMA5"},"Id":"00QR000000JNXLTMA5","Name":"Armanda
+        Bogan","FirstName":"Armanda","LastName":"Bogan","Salutation":null,"Subject__c":null,"Company":"not
         yet known","Phone":null,"Website":null,"Status":"Converted","Email":"f@f.com","LeadSource":"OSC
         Faculty","Newsletter__c":true,"Newsletter_Opt_In__c":true,"Adoption_Status__c":"High
-        Interest in Adopting","Number_of_Students__c":30.0,"OS_Accounts_ID__c":"1","accounts_uuid_c__c":null,"Application_Source__c":"Tutor
-        Signup","Role__c":"Faculty"}]}'
+        Interest in Adopting","Number_of_Students__c":30.0,"OS_Accounts_ID__c":"1","accounts_uuid_c__c":"ac80c504-7001-434b-94db-b25bafcef0d1","Application_Source__c":"Tutor
+        Signup","Role__c":"Faculty","who_chooses_books__c":null,"FV_Final__c":false}]}'
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 05:48:45 GMT
+  recorded_at: Fri, 03 Jul 2020 01:14:11 GMT
 recorded_with: VCR 3.0.3
