+++ spec/cassettes/Newflow_EducatorSignup_VerifyEducator/sf_setup.yml
@@ -21,7 +21,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Tue, 23 Jun 2020 05:48:40 GMT
+      - Thu, 02 Jul 2020 18:57:27 GMT
       Strict-Transport-Security:
       - max-age=31536000; includeSubDomains
       X-Content-Type-Options:
@@ -29,16 +29,16 @@ http_interactions:
       X-Xss-Protection:
       - 1; mode=block
       X-B3-Traceid:
-      - 1a3eb7ba9b06153c
+      - 74dda48844f16ae6
       X-B3-Spanid:
-      - 1a3eb7ba9b06153c
+      - 74dda48844f16ae6
       X-B3-Sampled:
       - '0'
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=MLlEmLUVEeq6QM_MnAQl9Q; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 05:48:40 GMT; Max-Age=31536000
+      - BrowserId=36YFr7yVEeqOFONoxY_gFg; domain=.salesforce.com; path=/; expires=Fri,
+        02-Jul-2021 18:57:27 GMT; Max-Age=31536000
       Expires:
       - Thu, 01 Jan 1970 00:00:00 GMT
       X-Readonlymode:
@@ -51,7 +51,7 @@ http_interactions:
       - chunked
     body:
       encoding: ASCII-8BIT
-      string: '{"access_token":"<salesforce_access_token>","instance_url":"<salesforce_instance_url>","id":"<salesforce_id>","token_type":"Bearer","issued_at":"1592891320673","signature":"<salesforce_signature>"}'
+      string: '{"access_token":"<salesforce_access_token>","instance_url":"<salesforce_instance_url>","id":"<salesforce_id>","token_type":"Bearer","issued_at":"1593716248050","signature":"<salesforce_signature>"}'
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 05:48:40 GMT
+  recorded_at: Thu, 02 Jul 2020 18:57:28 GMT
 recorded_with: VCR 3.0.3
