+++ spec/cassettes/Newflow_EducatorSignup_VerifyEducator/when_success/works.yml
@@ -25,7 +25,7 @@ http_interactions:
       message: Created
     headers:
       Date:
-      - Tue, 23 Jun 2020 20:02:45 GMT
+      - Thu, 02 Jul 2020 18:57:29 GMT
       Strict-Transport-Security:
       - max-age=31536002; includeSubDomains
       Public-Key-Pins-Report-Only:
@@ -37,20 +37,20 @@ http_interactions:
       X-Robots-Tag:
       - none
       X-B3-Traceid:
-      - 709a45cfbc9cd182
+      - e33252019c3a5787
       X-B3-Spanid:
-      - 709a45cfbc9cd182
+      - e33252019c3a5787
       X-B3-Sampled:
       - '0'
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=gOQO-bWMEeqhr7XRYgl8gg; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 20:02:45 GMT; Max-Age=31536000
+      - BrowserId=4JrGSryVEeqmxjvGPZToEg; domain=.salesforce.com; path=/; expires=Fri,
+        02-Jul-2021 18:57:29 GMT; Max-Age=31536000
       Sforce-Limit-Info:
-      - api-usage=5468/51000
+      - api-usage=2721/51000
       Location:
-      - "/services/data/v37.0/sobjects/Lead/00QR000000J6OpeMAF"
+      - "/services/data/v37.0/sobjects/Lead/00QR000000JNXKkMAP"
       Content-Type:
       - application/json;charset=UTF-8
       Vary:
@@ -59,12 +59,12 @@ http_interactions:
       - chunked
     body:
       encoding: ASCII-8BIT
-      string: '{"id":"00QR000000J6OpeMAF","success":true,"errors":[]}'
+      string: '{"id":"00QR000000JNXKkMAP","success":true,"errors":[]}'
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 20:02:47 GMT
+  recorded_at: Thu, 02 Jul 2020 18:57:32 GMT
 - request:
     method: get
-    uri: "<salesforce_instance_url_lower>/services/data/v37.0/query?q=SELECT%20Id,%20Name,%20FirstName,%20LastName,%20Salutation,%20Subject__c,%20Company,%20Phone,%20Website,%20Status,%20Email,%20LeadSource,%20Newsletter__c,%20Newsletter_Opt_In__c,%20Adoption_Status__c,%20Number_of_Students__c,%20OS_Accounts_ID__c,%20accounts_uuid_c__c,%20Application_Source__c,%20Role__c%20FROM%20Lead%20WHERE%20(Id%20=%20%2700QR000000J6OpeMAF%27)%20LIMIT%201"
+    uri: "<salesforce_instance_url_lower>/services/data/v37.0/query?q=SELECT%20Id,%20Name,%20FirstName,%20LastName,%20Salutation,%20Subject__c,%20Company,%20Phone,%20Website,%20Status,%20Email,%20LeadSource,%20Newsletter__c,%20Newsletter_Opt_In__c,%20Adoption_Status__c,%20Number_of_Students__c,%20OS_Accounts_ID__c,%20accounts_uuid_c__c,%20Application_Source__c,%20Role__c%20FROM%20Lead%20WHERE%20(Id%20=%20%2700QR000000JNXKkMAP%27)%20LIMIT%201"
     body:
       encoding: US-ASCII
       string: ''
@@ -83,7 +83,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Tue, 23 Jun 2020 20:02:47 GMT
+      - Thu, 02 Jul 2020 18:57:32 GMT
       Strict-Transport-Security:
       - max-age=31536002; includeSubDomains
       Public-Key-Pins-Report-Only:
@@ -95,18 +95,18 @@ http_interactions:
       X-Robots-Tag:
       - none
       X-B3-Traceid:
-      - c7347e2ad32af34c
+      - 123cc7b43244d3dc
       X-B3-Spanid:
-      - c7347e2ad32af34c
+      - 123cc7b43244d3dc
       X-B3-Sampled:
       - '0'
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=gnsqU7WMEequPcEfqIP_TA; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 20:02:47 GMT; Max-Age=31536000
+      - BrowserId=4qq7CryVEeqpD0ETOFi64g; domain=.salesforce.com; path=/; expires=Fri,
+        02-Jul-2021 18:57:32 GMT; Max-Age=31536000
       Sforce-Limit-Info:
-      - api-usage=5468/51000
+      - api-usage=2721/51000
       Content-Type:
       - application/json;charset=UTF-8
       Vary:
@@ -115,21 +115,21 @@ http_interactions:
       - chunked
     body:
       encoding: ASCII-8BIT
-      string: '{"totalSize":1,"done":true,"records":[{"attributes":{"type":"Lead","url":"/services/data/v37.0/sobjects/Lead/00QR000000J6OpeMAF"},"Id":"00QR000000J6OpeMAF","Name":"Bryan
+      string: '{"totalSize":1,"done":true,"records":[{"attributes":{"type":"Lead","url":"/services/data/v37.0/sobjects/Lead/00QR000000JNXKkMAP"},"Id":"00QR000000JNXKkMAP","Name":"Bryan
         Eli","FirstName":"Bryan","LastName":"Eli","Salutation":null,"Subject__c":null,"Company":"not
         yet known","Phone":null,"Website":null,"Status":"Converted","Email":"f@f.com","LeadSource":"OSC
         Faculty","Newsletter__c":false,"Newsletter_Opt_In__c":false,"Adoption_Status__c":"High
         Interest in Adopting","Number_of_Students__c":30.0,"OS_Accounts_ID__c":"1","accounts_uuid_c__c":null,"Application_Source__c":"Tutor
         Signup","Role__c":"Faculty"}]}'
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 20:02:47 GMT
+  recorded_at: Thu, 02 Jul 2020 18:57:32 GMT
 - request:
     method: patch
-    uri: "<salesforce_instance_url_lower>/services/data/v37.0/sobjects/Lead/00QR000000J6OpeMAF"
+    uri: "<salesforce_instance_url_lower>/services/data/v37.0/sobjects/Lead/00QR000000JNXKkMAP"
     body:
       encoding: UTF-8
-      string: '{"FirstName":"Grisel","LastName":"Mraz","Company":"Jacobi, Harris and
-        Nolan"}'
+      string: '{"FirstName":"Elbert","LastName":"Becker","Company":"Trantow, Spencer
+        and Kreiger","Role__c":"instructor"}'
     headers:
       User-Agent:
       - Faraday v0.17.3
@@ -147,7 +147,7 @@ http_interactions:
       message: No Content
     headers:
       Date:
-      - Tue, 23 Jun 2020 20:02:48 GMT
+      - Thu, 02 Jul 2020 18:57:33 GMT
       Strict-Transport-Security:
       - max-age=31536002; includeSubDomains
       Public-Key-Pins-Report-Only:
@@ -159,21 +159,21 @@ http_interactions:
       X-Robots-Tag:
       - none
       X-B3-Traceid:
-      - 6b570210824d6af5
+      - 9fc077331f4e7198
       X-B3-Spanid:
-      - 6b570210824d6af5
+      - 9fc077331f4e7198
       X-B3-Sampled:
       - '0'
       Cache-Control:
       - no-cache,must-revalidate,max-age=0,no-store,private
       Set-Cookie:
-      - BrowserId=gqcbyLWMEeqQ75Nv5YSsog; domain=.salesforce.com; path=/; expires=Wed,
-        23-Jun-2021 20:02:48 GMT; Max-Age=31536000
+      - BrowserId=4ubZRLyVEeqFmrnUwSpwPw; domain=.salesforce.com; path=/; expires=Fri,
+        02-Jul-2021 18:57:33 GMT; Max-Age=31536000
       Sforce-Limit-Info:
-      - api-usage=5468/51000
+      - api-usage=2721/51000
     body:
       encoding: UTF-8
       string: ''
     http_version: 
-  recorded_at: Tue, 23 Jun 2020 20:02:48 GMT
+  recorded_at: Thu, 02 Jul 2020 18:57:33 GMT
 recorded_with: VCR 3.0.3
